% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shares_all_classes.R
\name{calc_shares_all_classes}
\alias{calc_shares_all_classes}
\title{Calculate the number of shares across all equity asset classes}
\usage{
calc_shares_all_classes(
  shares_outstanding,
  asset_type_id,
  include_types = c("CP", "EQ", "PF")
)
}
\arguments{
\item{shares_outstanding}{(numeric vector) contains the number of shares
outstanding}

\item{asset_type_id}{(character vector) contains an the asset type identifier}

\item{include_types}{(character vector) defines which of the \code{asset_type_id}
values should be included in the sum}
}
\value{
a numeric value containing the number of shares across all equity
asset classes
}
\description{
\code{calc_shares_all_classes()} calculates the number of shares
across all equity asset classes. The asset class is defined by the value in
\code{asset_type_id}, and only asset classes included in \code{include_types} are
summed.

\code{calc_shares_all_classes()} assumes that all input values are from the same
entity/company, so a single summed value is returned. If it is desired to
calculate the shares in all classes for multiple entities at the same time,
a typical usage would be to apply this function "by" some entity id, e.g.

\code{mutate(data, shares_all_classes = calc_shares_all_classes(shares_outstanding, asset_type_id), .by = "entity_id")}

For more details see \code{vignette("share_ownership_methodology")}.
}
