% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_abcd_scen_connection.R
\name{dataprep_abcd_scen_connection}
\alias{dataprep_abcd_scen_connection}
\title{Combine ABCD and scenario data into the \verb{[equity/bonds]_abcd_scenario.rds}
format that is used by portfolio.analysis}
\usage{
dataprep_abcd_scen_connection(
  abcd_data,
  scenario_data,
  reference_year,
  relevant_years,
  tech_exclude,
  scenario_geographies_list,
  sector_list,
  other_sector_list,
  global_aggregate_scenario_sources_list,
  global_aggregate_sector_list,
  scenario_regions,
  index_regions
)
}
\arguments{
\item{abcd_data}{A tibble containing the ABCD data}

\item{scenario_data}{A tibble containing the scenario data}

\item{reference_year}{A single numeric specifying the market share target
reference year}

\item{relevant_years}{A numeric vector containing all relevant years to be
calculated}

\item{tech_exclude}{A character vector containing the technologies to be
excluded}

\item{scenario_geographies_list}{A character vector containing the scenario
geographies to be used}

\item{sector_list}{A character vector containing the sectors to be included}

\item{other_sector_list}{A character vector containing the sectors considered
"other"}

\item{global_aggregate_scenario_sources_list}{A character vector containing
the scenario sources to be included in the global aggreagte}

\item{global_aggregate_sector_list}{A character vector containing the sectors
to be included in the global aggregate}

\item{scenario_regions}{A character vector containing the scenario regions}

\item{index_regions}{A character vector containing the index regions}
}
\value{
A tibble with the combined ABCD and scenario data
}
\description{
Combine ABCD and scenario data into the \verb{[equity/bonds]_abcd_scenario.rds}
format that is used by portfolio.analysis
}
