% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_relevant_years.R
\name{determine_relevant_years}
\alias{determine_relevant_years}
\title{Determine relevant years}
\usage{
determine_relevant_years(market_share_target_reference_year, time_horizon)
}
\arguments{
\item{market_share_target_reference_year}{A single numeric value determining
the Market Share target reference year}

\item{time_horizon}{A single numeric value determining the number of forward
looking years}
}
\value{
A numeric vector containg all of the relevant years
}
\description{
Determine relevant years
}
