% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ar_advanced_company_indicators.R
\name{import_ar_masterdata}
\alias{import_ar_masterdata}
\title{Import the data from a version of Asset Resolution's bespoke
\code{masterdata_*.csv} files into a tidy data frame.}
\usage{
import_ar_masterdata(filepath, drop_nas = TRUE, id_as_string = FALSE)
}
\arguments{
\item{filepath}{Path to the CSV file.}

\item{drop_nas}{A logical indicating whether rows with an \code{NA} value after
pivoting to long-format should be dropped (deafult is \code{TRUE}).}

\item{id_as_string}{A logical indicating whether the \code{company_id} column
should be imported as a character vector. By default, the \code{company_id}
column is imported as a numeric (i.e. \code{FALSE}).}
}
\value{
A tidy, long-format tibble of all the data in the \code{masterdata_*.csv}
file with an added \code{consolidation_method} column to record which of
"ownership" or "debt" file was imported.
}
\description{
Import the data from a version of Asset Resolution's bespoke
\code{masterdata_*.csv} files into a tidy data frame.
}
