% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_ar_company_id__credit_parent_ar_company_id.R
\name{prepare_ar_company_id__credit_parent_ar_company_id}
\alias{prepare_ar_company_id__credit_parent_ar_company_id}
\title{Prepare an \code{ar_company_id__credit_parent_ar_company_id} lookup table from the
\code{entity_info} data}
\usage{
prepare_ar_company_id__credit_parent_ar_company_id(entity_info)
}
\arguments{
\item{entity_info}{A data frame containing the entity info}
}
\value{
A tibble
}
\description{
Prepare an \code{ar_company_id__credit_parent_ar_company_id} lookup table from the
\code{entity_info} data
}
