% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_ar_company_id__sectors_with_assets__debt.R
\name{prepare_ar_company_id__sectors_with_assets__debt}
\alias{prepare_ar_company_id__sectors_with_assets__debt}
\title{Prepare an \code{ar_company_id__sectors_with_assets__debt} lookup table from the
\code{masterdata_debt_datastore} data}
\usage{
prepare_ar_company_id__sectors_with_assets__debt(
  masterdata_debt_datastore,
  relevant_years
)
}
\arguments{
\item{masterdata_debt_datastore}{A data frame containing processed
production data from Asset Impact's masterdata_ownership CSV}

\item{relevant_years}{A numeric vector containing the relevant years of data
to include}
}
\value{
A tibble
}
\description{
Prepare an \code{ar_company_id__sectors_with_assets__debt} lookup table from the
\code{masterdata_debt_datastore} data
}
