% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_ar_company_id__sectors_with_assets__ownership.R
\name{prepare_ar_company_id__sectors_with_assets__ownership}
\alias{prepare_ar_company_id__sectors_with_assets__ownership}
\title{Prepare an \code{ar_company_id__sectors_with_assets__ownership} lookup table from
the \code{masterdata_ownership_datastore} data}
\usage{
prepare_ar_company_id__sectors_with_assets__ownership(
  masterdata_ownership_datastore,
  relevant_years
)
}
\arguments{
\item{masterdata_ownership_datastore}{A data frame containing processed
production data from Asset Impact's masterdata_ownership CSV}

\item{relevant_years}{A numeric vector containing the relevant years of data
to include}
}
\value{
A tibble
}
\description{
Prepare an \code{ar_company_id__sectors_with_assets__ownership} lookup table from
the \code{masterdata_ownership_datastore} data
}
