% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_company_id__creditor_company_id.R
\name{prepare_company_id__creditor_company_id}
\alias{prepare_company_id__creditor_company_id}
\title{Prepare a \code{company_id__creditor_company_id} lookup table from Asset Impact's
\code{masterdata_debt} data}
\usage{
prepare_company_id__creditor_company_id(masterdata_debt)
}
\arguments{
\item{masterdata_debt}{A data frame containing raw production data from Asset
Impact's masterdata_debt CSV}
}
\value{
A tibble
}
\description{
Prepare a \code{company_id__creditor_company_id} lookup table from Asset Impact's
\code{masterdata_debt} data
}
