% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_entity_info.R
\name{prepare_entity_info}
\alias{prepare_entity_info}
\title{Prepare the \code{entity_info} output data frame from data frames imported from
the \code{factset_entity_info.rds} and \code{ar_company_id__factset_entity_id.rds}
files}
\usage{
prepare_entity_info(
  data,
  factset_entity_id__ar_company_id,
  factset_industry_map_bridge,
  factset_manual_pacta_sector_override
)
}
\arguments{
\item{data}{A data frame containing the imported \code{factset_entity_info.rds}
file}

\item{factset_entity_id__ar_company_id}{A data frame containing the imported
\code{ar_company_id__factset_entity_id.rds} file with columns
\code{factset_entity_id} and \code{ar_company_id}}

\item{factset_industry_map_bridge}{A data frame containing the imported
\code{factset_industry_map_bridge.rds}}

\item{factset_manual_pacta_sector_override}{A data frame containing the imported
\code{factset_manual_pacta_sector_override}}
}
\value{
A tibble properly prepared to be saved as the \code{entity_info.rds}
output file
}
\description{
Prepare the \code{entity_info} output data frame from data frames imported from
the \code{factset_entity_info.rds} and \code{ar_company_id__factset_entity_id.rds}
files
}
