% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_factset_entity_id__ar_company_id.R
\name{prepare_factset_entity_id__ar_company_id}
\alias{prepare_factset_entity_id__ar_company_id}
\title{Prepare a \code{factset_entity_id__ar_company_id} lookup table from Asset Impact's
\code{ar_company_id__factset_entity_id} crosswalk}
\usage{
prepare_factset_entity_id__ar_company_id(ar_company_id__factset_entity_id)
}
\arguments{
\item{ar_company_id__factset_entity_id}{A data frame containing a production
data company ID to financial data entity lookup table}
}
\value{
A tibble
}
\description{
Prepare a \code{factset_entity_id__ar_company_id} lookup table from Asset Impact's
\code{ar_company_id__factset_entity_id} crosswalk
}
