% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_factset_entity_id__credit_parent_id.R
\name{prepare_factset_entity_id__credit_parent_id}
\alias{prepare_factset_entity_id__credit_parent_id}
\title{Prepare a \code{factset_entity_id__credit_parent_id} lookup table from
\code{entity_info}}
\usage{
prepare_factset_entity_id__credit_parent_id(entity_info)
}
\arguments{
\item{entity_info}{A data frame containing the entity info}
}
\value{
A tibble
}
\description{
Prepare a \code{factset_entity_id__credit_parent_id} lookup table from
\code{entity_info}
}
