% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_factset_entity_id__security_mapped_sector.R
\name{prepare_factset_entity_id__security_mapped_sector}
\alias{prepare_factset_entity_id__security_mapped_sector}
\title{Prepare a \code{factset_entity_id__security_mapped_sector} lookup table from
\code{entity_info}}
\usage{
prepare_factset_entity_id__security_mapped_sector(entity_info)
}
\arguments{
\item{entity_info}{A data frame containing the entity info}
}
\value{
A tibble
}
\description{
Prepare a \code{factset_entity_id__security_mapped_sector} lookup table from
\code{entity_info}
}
