% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_financial_data.R
\name{prepare_financial_data}
\alias{prepare_financial_data}
\title{Prepare the \code{financial_data} output data frame from the imported
\code{factset_financial_data.rds} file}
\usage{
prepare_financial_data(data, issue_code_bridge)
}
\arguments{
\item{data}{A data frame containing the imported \code{factset_financial_data.rds}
file}

\item{issue_code_bridge}{A data frame containing data that bridges from
factset issue codes to one of \code{c("Listed Equity", "Corporate Bond", "Fund", "Other")}}
}
\value{
A tibble properly prepared to be saved as the \code{financial_data.rds}
output file
}
\description{
Prepare the \code{financial_data} output data frame from the imported
\code{factset_financial_data.rds} file
}
