% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fund_data.R
\name{prepare_fund_data}
\alias{prepare_fund_data}
\title{Prepare fund data, filtering to funds with data according to a given
threshold and adding a \code{MISSINGWEIGHT} holding for the difference}
\usage{
prepare_fund_data(fund_data, threshold = 0)
}
\arguments{
\item{fund_data}{A data frame containing fund data}

\item{threshold}{A numeric value between 0 and 1 (inclusive) indicating the
allowable percentage of the total fund value that the summed values of its
component holdings should be equal to or greater than}
}
\value{
A tibble
}
\description{
Prepare fund data, filtering to funds with data according to a given
threshold and adding a \code{MISSINGWEIGHT} holding for the difference
}
