% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isin_to_fund_table.R
\name{prepare_isin_to_fund_table}
\alias{prepare_isin_to_fund_table}
\title{Prepare \code{isin_to_fund_table}, filtering out fsyms that have more than 1 row
and either no fund data or fund data for both rows}
\usage{
prepare_isin_to_fund_table(isin_to_fund_table, fund_data)
}
\arguments{
\item{isin_to_fund_table}{A data frame containing isin_to_fund_table data}

\item{fund_data}{A data frame containing fund data}
}
\value{
A tibble
}
\description{
Prepare \code{isin_to_fund_table}, filtering out fsyms that have more than 1 row
and either no fund data or fund data for both rows
}
