% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_iss_company_emissions.R
\name{prepare_iss_company_emissions}
\alias{prepare_iss_company_emissions}
\title{Prepare an \code{iss_company_emissions} object from \code{factset_iss_emissions_data}}
\usage{
prepare_iss_company_emissions(factset_iss_emissions_data)
}
\arguments{
\item{factset_iss_emissions_data}{A data frame containing ISS emissions data}
}
\value{
A tibble
}
\description{
Prepare an \code{iss_company_emissions} object from \code{factset_iss_emissions_data}
}
