% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_iss_entity_emission_intensities.R
\name{prepare_iss_entity_emission_intensities}
\alias{prepare_iss_entity_emission_intensities}
\title{Prepare a \code{iss_entity_emission_intensities} object}
\usage{
prepare_iss_entity_emission_intensities(
  iss_company_emissions,
  factset_financial_data,
  factset_entity_info,
  factset_entity_financing_data,
  currencies
)
}
\arguments{
\item{iss_company_emissions}{A data frame containing \code{iss_company_emissions}
data}

\item{factset_financial_data}{A data frame containing
\code{factset_financial_data} data}

\item{factset_entity_info}{A data frame containing \code{factset_entity_info} data}

\item{factset_entity_financing_data}{A data frame containing
\code{factset_entity_financing_data} data}

\item{currencies}{A data frame containing currency exchange rate data}
}
\value{
A data frame containing the prepared
\code{iss_entity_emission_intensities} object
}
\description{
Prepare a \code{iss_entity_emission_intensities} object
}
