% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_masterdata.R
\name{prepare_masterdata}
\alias{prepare_masterdata}
\title{Prepare the \code{masterdata_ownership_datastore} or \code{masterdata_debt_datastore}
output data frame from an import of a raw AR masterdata_* CSV}
\usage{
prepare_masterdata(
  data,
  ar_company_id__country_of_domicile,
  pacta_financial_timestamp,
  zero_emission_factor_techs
)
}
\arguments{
\item{data}{A dataframe containing the raw input of an AR masterdata_* CSV
files}

\item{ar_company_id__country_of_domicile}{A data frame with two columns
mapping \code{ar_company_id} to \code{country_of_domicile}}

\item{pacta_financial_timestamp}{A single element character vector specifying
the timestamp in the PACTA format, e.g. "2021Q4"}

\item{zero_emission_factor_techs}{A character vector listing technologies
that will have emission factors manually forced to 0}
}
\value{
A tibble properly prepared to be saved as the
\code{masterdata_ownership_datastore.rds} or \code{masterdata_debt_datastore.rds}
output file
}
\description{
Prepare the \code{masterdata_ownership_datastore} or \code{masterdata_debt_datastore}
output data frame from an import of a raw AR masterdata_* CSV
}
