% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_masterdata_debt.R
\name{prepare_masterdata_debt}
\alias{prepare_masterdata_debt}
\title{Prepare the \code{masterdata_debt_datastore} object from a raw masterdata_debt CSV}
\usage{
prepare_masterdata_debt(
  masterdata_debt_raw,
  ar_company_id__country_of_domicile,
  ar_company_id__credit_parent_ar_company_id,
  pacta_financial_timestamp,
  zero_emission_factor_techs
)
}
\arguments{
\item{masterdata_debt_raw}{A data frame containing the raw data from a
masterdata_debt CSV}

\item{ar_company_id__country_of_domicile}{A data frame containing an
\code{ar_company_id} to \code{country_of_domicile} lookup}

\item{ar_company_id__credit_parent_ar_company_id}{A data frame containing an
\code{ar_company_id} to \code{credit_parent_ar_company_id} lookup}

\item{pacta_financial_timestamp}{A single character vector containing the
PACTA financial timestamp, e.g. \verb{2023Q4}}

\item{zero_emission_factor_techs}{A character vector containing the zero
emission factor technologies}
}
\value{
A data frame containing the prepared masterdata_debt_datastore
}
\description{
Prepare the \code{masterdata_debt_datastore} object from a raw masterdata_debt CSV
}
