% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_total_fund_list.R
\name{prepare_total_fund_list}
\alias{prepare_total_fund_list}
\title{Prepare a \code{total_fund_list} object from \code{fund_data}}
\usage{
prepare_total_fund_list(fund_data)
}
\arguments{
\item{fund_data}{A data frame containing fund data}
}
\value{
A tibble
}
\description{
Prepare a \code{total_fund_list} object from \code{fund_data}
}
