% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter_to_factset_timestamp.R
\name{quarter_to_factset_timestamp}
\alias{quarter_to_factset_timestamp}
\title{Convert a PACTA style quarter string to a FactSet style date string for the
last day of that quarter}
\usage{
quarter_to_factset_timestamp(quarter)
}
\arguments{
\item{quarter}{A character vector containing PACTA style quarter strings in
the form e.g. "2021Q4"}
}
\value{
A character vector containing the equivalent FactSet style date
strings for the last day of the quarter e.g. "2021-12-31"
}
\description{
Convert a PACTA style quarter string to a FactSet style date string for the
last day of that quarter
}
