% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter_to_imf_timestamp.R
\name{quarter_to_imf_timestamp}
\alias{quarter_to_imf_timestamp}
\title{Convert a PACTA style quarter string to an IMF style quarter string}
\usage{
quarter_to_imf_timestamp(quarter)
}
\arguments{
\item{quarter}{A character vector containing PACTA style quarter strings in
the form e.g. "2021Q4"X}
}
\value{
A character vector containing the equivalent IMF style quarter
strings e.g. "2021-Q4"
}
\description{
Convert a PACTA style quarter string to an IMF style quarter string
}
