% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_asset_type_names.R
\name{standardize_asset_type_names}
\alias{standardize_asset_type_names}
\title{Standardize asset type names}
\usage{
standardize_asset_type_names(factset_issue_code_bridge)
}
\arguments{
\item{factset_issue_code_bridge}{A data frame containing the FactSet issue
code bridge}
}
\value{
A data frame containing a FactSet \code{issue_type_code} to \code{asset_type}
lookup
}
\description{
Standardize asset type names
}
